add calculators and change views

+++ FitApp/Models/ONERM.cs
@@ -7,8 +7,8 @@ namespace FitApp.Models
 {
     public class ONERM
     {
-        //(weight*reps*0,0333) +weight
-        public int Weight { get; set; }
-        public int Reps { get; set; }
+        
+        public float Weight { get; set; }
+        public float Reps { get; set; }
     }
 }
